test(pod): split the engine tests under the line limit
